requirements.txt
duckdb==1.1.3
pandas==2.2.2
pyarrow==16.1.0
python-dotenv==1.0.1
tabulate==0.9.0
